The setValue method was moved to MobileElement.
